Refined core logic and sprinkled some magic
