+++ requirements.txt
@@ -9,3 +9,4 @@ requests==2.32.3
 beautifulsoup4==4.12.3
 tenacity==8.5.0
 selenium==4.25.0
+jinja2==3.1.4
